git-svn-id: file:///home/koke/ebox2git/svn/branches/maintenance-0.12-0.12.10/hardy-remoteservices@12292 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ ebox/src/EBox/Backup.pm
@@ -579,11 +579,16 @@ sub _ensureBackupdirExistence
 #
 # Parameters:
 #
-#                   description - backup's description (backwards compability mode)
-#                   fullBackup  - wether do a full backup or  backup only configuration (default: false)
+#                   description - backup's description (default: 'Backup')
+#                   fullBackup  - whether do a full backup or  backup only configuration (default: false)
+#                   bug         - whether this backup is intended for a bug report
+#                                 (one consequence of this is that we must clear
+#                                  private data) 
 #
-#  Returns:
-#    the progress indicator object whihc represents the progress of the restauration
+#                   remoteBackup -- whether this a backup intended to be a remote backup
+#
+# Returns:
+#     progress indicator object of the operation
 #
 # Exceptions:
 #	
@@ -593,23 +598,47 @@ sub prepareMakeBackup
 {
   my ($self, %options) = @_;
 
-  # make sure description is scaped
-  $options{description} = q{'} . $options{description} . q{'};
-  my @scriptParams = %options; 
+  my $scriptParams = '';
+
+  if ( $options{remoteBackup} ) {
+      $scriptParams .= ' --remote-backup ';
+      # Make sure remote backup name is scaped
+      $scriptParams .= q{'} . $options{remoteBackup} . q{'};
+  }
+
+  if (exists $options{description}) {
+      $scriptParams .= ' --description ';
+      # make sure description is scaped
+      $scriptParams .= q{'} . $options{description} . q{'};
+  }
+
+  if ($options{fullBackup}) {
+      $scriptParams .= ' --full-backup';
+  }
+  else {
+      $scriptParams .= ' --config-backup';
+  }
+
+  if ($options{bug}) {
+      $scriptParams .= ' --bug-report';
+  }
 
   my $makeBackupScript = EBox::Config::pkgdata() . 'ebox-make-backup';
-  $makeBackupScript    .= "  @scriptParams";
-  
+  $makeBackupScript    .= $scriptParams;
+
   my $global     = EBox::Global->getInstance();
-  my $totalTicks = scalar @{ $global->modNames } + 2; # there are one task for
+  my $totalTicks = scalar @{ $global->modNames() } + 2; # there are one task for
                                                       # each module plus two
                                                       # tasks for writing the
                                                       # archive  file
-           
 
-  my $progressIndicator =  EBox::ProgressIndicator->create(
+  my @progressIndicatorParams = (
 			     executable => $makeBackupScript,
 			     totalTicks => $totalTicks,
+						    );
+
+  my $progressIndicator =  EBox::ProgressIndicator->create(
+							@progressIndicatorParams
 						    );
 
   $progressIndicator->runExecutable();
@@ -628,12 +657,18 @@ sub prepareMakeBackup
 #                   progress-  progress indicator 
 #                       associated with this operation (optionak)
 #                   description - backup's description (default: 'Backup')
-#                   fullBackup  - wether do a full backup or  backup only configuration (default: false)
+#                   fullBackup  - whether do a full backup or  backup only configuration (default: false)
+#                   bug         - whether this backup is intended for a bug report
+#                                 (one consequence of this is that we must clear
+#                                  private data) 
+#
+#  Returns:
+#         - path to the new backup archive
 #
 # Exceptions:
-#	
-#	Internal - If backup fails
-#       Exteanl   - If modules have unsaved changes
+#
+#       Internal - If backup fails
+#       External - If modules have unsaved changes
 sub makeBackup # (options) 
 {
   my ($self, %options) = @_;
@@ -680,7 +715,7 @@ sub makeBackup # (options)
   my $backupFinalPath;
   try {
       $progress->notifyTick();
-      $progress->setMessage(__('Writing backup file to hard disc'));
+      $progress->setMessage(__('Writing backup file to hard disk'));
 
       $backupFinalPath = $self->_moveToArchives($filename, $backupdir);   
 
@@ -880,11 +915,13 @@ sub  _checkSize
 #
 # Parameters:
 #
-#       file - backup's file (as positional parameter) fullRestore - wether do a
-#       full restore or restore only configuration (default: false) 
-#
+#       file - backup's file (as positional parameter) 
+# fullRestore - wether do a full restore or restore only configuration (default: false)
+#       dataRestore - wether do a data-only restore
+#       forceDependencies - wether ignore dependency errors between modules
+#        deleteBackup      - deletes the backup after resroting it or if the process is aborted
 #  Returns:
-#    the progress indicator object whihc represents the progress of the restauration
+#    the progress indicator object which represents the progress of the restauration
 #
 # Exceptions:
 #	
@@ -907,6 +944,12 @@ sub prepareRestoreBackup
   if (exists $options{forceDependencies}) {
     if ($options{forceDependencies}) {
       $execOptions .= '--force-dependencies ';
+    }
+  }
+
+  if (exists $options{deleteBackup}) {
+    if ($options{deleteBackup}) {
+      $execOptions .= '--delete-backup ';
     }
   }
 
@@ -943,6 +986,9 @@ sub prepareRestoreBackup
 #       progressIndicator - Progress indicator associated
 #                       with htis operation (optional )
 # fullRestore - wether do a full restore or restore only configuration (default: false)
+#       dataRestore - wether do a data-only restore
+#       forceDependencies - wether ignore dependency errors between modules
+#        deleteBackup      - deletes the backup after resroting it or if the process is aborted
 #
 # Exceptions:
 #	
@@ -976,11 +1022,12 @@ sub restoreBackup # (file, %options)
 					      
 					  },
 			   forceDependencies => {default => 0 },
+			   deleteBackup      => { default => 0},
 			  }
 		);
 
-  _ensureBackupdirExistence();
-  $self->_checkSize($file);
+
+
 
   my $progress = $options{progress};
   if (not $progress) {
@@ -988,15 +1035,18 @@ sub restoreBackup # (file, %options)
     $options{progress} = $progress;
   }
 
-
-
   EBox::debug("restore backup id: " . $progress->id);
   $progress->started or
     throw EBox::Exceptions::Internal("ProgressIndicator's executable has not been run");
 
-  my $tempdir = $self->_unpackAndVerify($file, $options{fullRestore});
-
+  my $tempdir;
   try {
+    _ensureBackupdirExistence();
+
+    $self->_checkSize($file);
+
+    $tempdir = $self->_unpackAndVerify($file, $options{fullRestore});
+
     $self->_unpackModulesRestoreData($tempdir);
 
     my @modules  = @{ $self->_modInstancesForRestore($file, %options) };
@@ -1027,7 +1077,13 @@ sub restoreBackup # (file, %options)
     $progress->setAsFinished(); 
   }
   finally {
+    if ($tempdir) {
       system 'rm -rf $tempdir';
+    }
+    if ($options{deleteBackup}) {
+      unlink $file;
+    }
+
   };
 }
